+++ android/app/src/main/java/com/aipx/eva/MainActivity.java
@@ -1,0 +1,5 @@
+package com.aipx.eva;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}